refactor(locate): abstract POI spatial search to locator helper
- Extract POI grid and distance math from `locate` command to `find_nearby_pois` in `locator`
- Clean up `PoiExecutor` implementation to match structure and biome executor patterns
